Move common (duplicateD) _run_pre_inference to parent PySHACL runner.
